requirement.txt
websocket-client
TA-Lib
python-binance
numpy
